+++ protocol/x/vault/keeper/msg_server_withdraw_from_vault.go
@@ -6,10 +6,23 @@ import (
 	"github.com/dydxprotocol/v4-chain/protocol/x/vault/types"
 )
 
-// TODO: implement
+// WithdrawFromVault withdraws from a vault to a subaccount.
 func (k msgServer) WithdrawFromVault(
 	goCtx context.Context,
 	msg *types.MsgWithdrawFromVault,
 ) (*types.MsgWithdrawFromVaultResponse, error) {
+	// TODO(TRA-461): validate
+
+	// TODO(TRA-462): Calculate effective amount to withdraw + shares to redeem with slippage and user equity.
+
+	// TODO(TRA-461): Redeem shares for the vault.
+
+	// TODO(TRA-461): Transfer asset from vault to recipient subaccount.
+	// should transfer happen after redeeming shares? why?
+
+	// TODO(TRA-461): emit metric on vault equity.
+
+	// TODO(TRA-461):: Get info on shares after the withdrawal.
+
 	return &types.MsgWithdrawFromVaultResponse{}, nil
 }
