+++ protocol/x/vault/keeper/msg_server_withdraw_from_vault_test.go
@@ -1,0 +1,3 @@
+package keeper_test
+
+// TODO(TRA-461): add tests
