+++ protocol/x/vault/types/msg_withdraw_from_vault.go
@@ -5,15 +5,15 @@ import (
 	"github.com/cosmos/cosmos-sdk/types"
 )
 
-var _ types.Msg = &MsgDepositToVault{}
+var _ types.Msg = &MsgWithdrawFromVault{}
 
-// ValidateBasic performs stateless validation on a MsgDepositToVault.
-func (msg *MsgDepositToVault) ValidateBasic() error {
+// ValidateBasic performs stateless validation on a MsgWithdrawFromVault.
+func (msg *MsgWithdrawFromVault) ValidateBasic() error {
 	// Note: msg signer must be the owner of the subaccount.
 	// This is enforced by the following notatino on the msg proto:
 	//    option (cosmos.msg.v1.signer) = "subaccount_id"
 
-	// Validate subaccount to deposit from.
+	// Validate subaccount to withdraw to.
 	if err := msg.SubaccountId.Validate(); err != nil {
 		return err
 	}
@@ -21,7 +21,7 @@ func (msg *MsgDepositToVault) ValidateBasic() error {
 	// Validate that quote quantums is positive and an uint64.
 	quoteQuantums := msg.QuoteQuantums.BigInt()
 	if quoteQuantums.Sign() <= 0 || !quoteQuantums.IsUint64() {
-		return errors.Wrap(ErrInvalidDepositAmount, "quote quantums must be strictly positive and less than 2^64")
+		return errors.Wrap(ErrInvalidWithdrawalAmount, "quote quantums must be strictly positive and less than 2^64")
 	}
 
 	return nil
