+++ protocol/x/vault/types/msg_withdraw_from_vault_test.go
@@ -12,27 +12,27 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestMsgDepositToVault_ValidateBasic(t *testing.T) {
+func TestMsgWithdrawFromVault_ValidateBasic(t *testing.T) {
 	tests := map[string]struct {
-		msg         types.MsgDepositToVault
+		msg         types.MsgWithdrawFromVault
 		expectedErr string
 	}{
 		"Success": {
-			msg: types.MsgDepositToVault{
+			msg: types.MsgWithdrawFromVault{
 				VaultId:       &constants.Vault_Clob0,
 				SubaccountId:  &constants.Alice_Num0,
 				QuoteQuantums: dtypes.NewInt(1),
 			},
 		},
 		"Success: max uint64 quote quantums": {
-			msg: types.MsgDepositToVault{
+			msg: types.MsgWithdrawFromVault{
 				VaultId:       &constants.Vault_Clob0,
 				SubaccountId:  &constants.Alice_Num0,
 				QuoteQuantums: dtypes.NewIntFromUint64(math.MaxUint64),
 			},
 		},
 		"Failure: quote quantums greater than max uint64": {
-			msg: types.MsgDepositToVault{
+			msg: types.MsgWithdrawFromVault{
 				VaultId:      &constants.Vault_Clob0,
 				SubaccountId: &constants.Alice_Num0,
 				QuoteQuantums: dtypes.NewIntFromBigInt(
@@ -42,26 +42,26 @@ func TestMsgDepositToVault_ValidateBasic(t *testing.T) {
 					),
 				),
 			},
-			expectedErr: "Deposit amount is invalid",
+			expectedErr: "Withdrawal amount is invalid",
 		},
 		"Failure: zero quote quantums": {
-			msg: types.MsgDepositToVault{
+			msg: types.MsgWithdrawFromVault{
 				VaultId:       &constants.Vault_Clob0,
 				SubaccountId:  &constants.Alice_Num0,
 				QuoteQuantums: dtypes.NewInt(0),
 			},
-			expectedErr: "Deposit amount is invalid",
+			expectedErr: "Withdrawal amount is invalid",
 		},
 		"Failure: negative quote quantums": {
-			msg: types.MsgDepositToVault{
+			msg: types.MsgWithdrawFromVault{
 				VaultId:       &constants.Vault_Clob0,
 				SubaccountId:  &constants.Alice_Num0,
 				QuoteQuantums: dtypes.NewInt(-1),
 			},
-			expectedErr: "Deposit amount is invalid",
+			expectedErr: "Withdrawal amount is invalid",
 		},
 		"Failure: invalid subaccount owner": {
-			msg: types.MsgDepositToVault{
+			msg: types.MsgWithdrawFromVault{
 				VaultId: &constants.Vault_Clob0,
 				SubaccountId: &satypes.SubaccountId{
 					Owner:  "invalid-owner",
